+++ src/library/Hasql/Pool.hs
@@ -85,7 +85,7 @@ acquire config = do
     join . atomically $ do
       entries <- flushTQueue connectionQueue
       let (agedEntries, unagedEntries) = partition (entryIsAged agingTimeoutNanos now) entries
-          (idleEntries, liveEntries) = partition (entryIsIdle agingTimeoutNanos now) unagedEntries
+          (idleEntries, liveEntries) = partition (entryIsIdle maxIdletimeNanos now) unagedEntries
       traverse_ (writeTQueue connectionQueue) liveEntries
       return $ do
         forM_ agedEntries $ \entry -> do